+++ es-app/src/guis/GuiMenu.cpp
@@ -304,18 +304,20 @@ GuiMenu::GuiMenu(Window *window) : GuiComponent(window), mMenu(window, "MAIN MEN
                          // For each activated system
                          std::vector<SystemData *> systems = SystemData::sSystemVector;
                          for (auto system = systems.begin(); system != systems.end(); system++) {
-                             ComponentListRow systemRow;
-                             auto systemText = std::make_shared<TextComponent>(mWindow, (*system)->getFullName(),
-                                                                               Font::get(FONT_SIZE_MEDIUM),
-                                                                               0x777777FF);
-                             auto bracket = makeArrow(mWindow);
-                             systemRow.addElement(systemText, true);
-                             systemRow.addElement(bracket, false);
-                             SystemData *systemData = (*system);
-                             systemRow.makeAcceptInputHandler([this, systemData] {
-                                 popSystemConfigurationGui(systemData, "");
-                             });
-                             configuration->addRow(systemRow);
+                             if ((*system) != SystemData::getFavoriteSystem()) {
+                                 ComponentListRow systemRow;
+                                 auto systemText = std::make_shared<TextComponent>(mWindow, (*system)->getFullName(),
+                                                                                   Font::get(FONT_SIZE_MEDIUM),
+                                                                                   0x777777FF);
+                                 auto bracket = makeArrow(mWindow);
+                                 systemRow.addElement(systemText, true);
+                                 systemRow.addElement(bracket, false);
+                                 SystemData *systemData = (*system);
+                                 systemRow.makeAcceptInputHandler([this, systemData] {
+                                     popSystemConfigurationGui(systemData, "");
+                                 });
+                                 configuration->addRow(systemRow);
+                             }
                          }
                          mWindow->pushGui(configuration);